refactor: split AI assistant bubble component into smaller components
