+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AssistantBubbleActionIcons.module.css
@@ -1,0 +1,8 @@
+.feedbackInput {
+    background-color: rgba(255, 255, 255, 0.3) !important;
+
+    --input-bd-focus: var(--mantine-color-gray-7);
+    &:focus {
+        outline: 4px solid var(--mantine-color-gray-2);
+    }
+}
